requirements.txt
backports.functools-lru-cache>=1.5
beautifulsoup4>=4.7.1
bs4>=0.0.1
cachetools>=3.1.1
cryptography>=1.7.1
enum34>=1.1.6
google-api-python-client>=1.7.4
google-auth>=1.5.1
google-auth-httplib2>=0.0.3
httplib2>=0.11.3
idna>=2.2
ipaddress>=1.0.17
keyring>=10.1
keyrings.alt>=1.3
ldap3>=2.6
mysql-connector-python>=8.0.16
oauth2client>=4.1.2
protobuf>=3.8.0
pyasn1>=0.4.5
pyasn1-modules>=0.2.5
pycrypto>=2.6.1
pygobject>=3.22.0
python-dateutil>=2.8.0
python-ldap>=3.2.0
pyxdg>=0.25
rsa>=4.0
SecretStorage>=2.3.1
six>=1.12.0
soupsieve>=1.9.2
uritemplate>=3.0.0
